2036: separate the TCAA functionality from the TCA (tca.go -> tcaa.go)
