+++ membersrvc/ca/tca.go
@@ -75,11 +75,6 @@ type TCAP struct {
 	tca *TCA
 }
 
-// TCAA serves the administrator GRPC interface of the TCA.
-type TCAA struct {
-	tca *TCA
-}
-
 // TCertSet contains relevant information of a set of tcerts
 type TCertSet struct {
 	Ts           int64
@@ -589,27 +584,6 @@ func (tcap *TCAP) RevokeCertificate(context.Context, *pb.TCertRevokeReq) (*pb.CA
 // RevokeCertificateSet revokes a certificate set from the TCA.  Not yet implemented.
 func (tcap *TCAP) RevokeCertificateSet(context.Context, *pb.TCertRevokeSetReq) (*pb.CAStatus, error) {
 	Trace.Println("grpc TCAP:RevokeCertificateSet")
-
-	return nil, errors.New("not yet implemented")
-}
-
-// RevokeCertificate revokes a certificate from the TCA.  Not yet implemented.
-func (tcaa *TCAA) RevokeCertificate(context.Context, *pb.TCertRevokeReq) (*pb.CAStatus, error) {
-	Trace.Println("grpc TCAA:RevokeCertificate")
-
-	return nil, errors.New("not yet implemented")
-}
-
-// RevokeCertificateSet revokes a certificate set from the TCA.  Not yet implemented.
-func (tcaa *TCAA) RevokeCertificateSet(context.Context, *pb.TCertRevokeSetReq) (*pb.CAStatus, error) {
-	Trace.Println("grpc TCAA:RevokeCertificateSet")
-
-	return nil, errors.New("not yet implemented")
-}
-
-// PublishCRL requests the creation of a certificate revocation list from the TCA.  Not yet implemented.
-func (tcaa *TCAA) PublishCRL(context.Context, *pb.TCertCRLReq) (*pb.CAStatus, error) {
-	Trace.Println("grpc TCAA:CreateCRL")
 
 	return nil, errors.New("not yet implemented")
 }
