+++ membersrvc/ca/tcaa.go
@@ -1,0 +1,50 @@
+/*
+Copyright IBM Corp. 2016 All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+		 http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package ca
+
+import (
+	"errors"
+
+	pb "github.com/hyperledger/fabric/membersrvc/protos"
+	"golang.org/x/net/context"
+)
+
+// TCAA serves the administrator GRPC interface of the TCA.
+type TCAA struct {
+	tca *TCA
+}
+
+// RevokeCertificate revokes a certificate from the TCA.  Not yet implemented.
+func (tcaa *TCAA) RevokeCertificate(context.Context, *pb.TCertRevokeReq) (*pb.CAStatus, error) {
+	Trace.Println("grpc TCAA:RevokeCertificate")
+
+	return nil, errors.New("not yet implemented")
+}
+
+// RevokeCertificateSet revokes a certificate set from the TCA.  Not yet implemented.
+func (tcaa *TCAA) RevokeCertificateSet(context.Context, *pb.TCertRevokeSetReq) (*pb.CAStatus, error) {
+	Trace.Println("grpc TCAA:RevokeCertificateSet")
+
+	return nil, errors.New("not yet implemented")
+}
+
+// PublishCRL requests the creation of a certificate revocation list from the TCA.  Not yet implemented.
+func (tcaa *TCAA) PublishCRL(context.Context, *pb.TCertCRLReq) (*pb.CAStatus, error) {
+	Trace.Println("grpc TCAA:CreateCRL")
+
+	return nil, errors.New("not yet implemented")
+}
